Extract CSA teacher LSE kernels
